Group field-related types
