+++ .changeset/pink-beds-rest.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/types': patch
+---
+
+Moved field-related types into their own internal module.
