+++ packages-next/types/src/config/lists.ts
@@ -1,9 +1,7 @@
-import { PrismaFieldAdapter } from '@keystone-next/adapter-prisma-legacy';
-import { Implementation } from '@keystone-next/fields';
-import { AdminMetaRootVal } from '../admin-meta';
-import type { BaseGeneratedListTypes, MaybePromise, JSONValue } from '../utils';
+import type { BaseGeneratedListTypes, MaybePromise } from '../utils';
 import type { ListHooks } from './hooks';
-import type { ListAccessControl, FieldAccessControl } from './access-control';
+import type { ListAccessControl } from './access-control';
+import type { FieldType, BaseFields } from './fields';
 
 export type ListSchemaConfig = Record<string, ListConfig<BaseGeneratedListTypes, any>>;
 
@@ -156,50 +154,6 @@ export type ListAdminUIConfig<
      * @default 'edit'
      */
     defaultFieldMode?: MaybeItemFunction<'edit' | 'read' | 'hidden'>;
-  };
-};
-
-export type BaseFields<TGeneratedListTypes extends BaseGeneratedListTypes> = {
-  [key: string]: FieldType<TGeneratedListTypes>;
-};
-
-export type FieldType<TGeneratedListTypes extends BaseGeneratedListTypes> = {
-  /**
-   * The real keystone type for the field
-   */
-  type: {
-    type: string;
-    implementation: typeof Implementation;
-    adapter: typeof PrismaFieldAdapter;
-    isRelationship?: boolean;
-  };
-  /**
-   * The config for the field
-   */
-  config: FieldConfig<TGeneratedListTypes>;
-  /**
-   * The resolved path to the views for the field type
-   */
-  views: string;
-  getAdminMeta?: (listKey: string, path: string, adminMeta: AdminMetaRootVal) => JSONValue;
-};
-
-export type FieldConfig<TGeneratedListTypes extends BaseGeneratedListTypes> = {
-  access?: FieldAccessControl<TGeneratedListTypes>;
-  hooks?: ListHooks<TGeneratedListTypes>; // really? ListHooks?
-  label?: string;
-  ui?: {
-    views?: string;
-    description?: string;
-    createView?: {
-      fieldMode?: MaybeSessionFunction<'edit' | 'hidden'>;
-    };
-    listView?: {
-      fieldMode?: MaybeSessionFunction<'read' | 'hidden'>;
-    };
-    itemView?: {
-      fieldMode?: MaybeItemFunction<'edit' | 'read' | 'hidden'>;
-    };
   };
 };
